refactor[api-server]: TEST_UID 하나로 통일되도록 수정
