+++ src/test/java/com/watermelon/server/Constants.java
@@ -2,7 +2,6 @@ package com.watermelon.server;
 
 public class Constants {
 
-    public static final String TEST_UID = "uid";
     public static final String TEST_NOT_UID = "not_uid";
     public static final String TEST_NAME = "name";
     public static final String TEST_ADDRESS = "address";
