+++ src/test/java/com/watermelon/server/MockLoginInterceptorConfig.java
@@ -8,8 +8,7 @@ import org.mockito.Mockito;
 import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.context.annotation.Bean;
 
-import static com.watermelon.server.Constants.TEST_TOKEN;
-import static com.watermelon.server.Constants.TEST_UID;
+import static com.watermelon.server.event.lottery.auth.service.TestTokenVerifier.TEST_UID;
 
 @TestConfiguration
 public class MockLoginInterceptorConfig {
