+++ src/test/java/com/watermelon/server/admin/service/AdminUserServiceTest.java
@@ -12,7 +12,6 @@ import org.mockito.Mockito;
 import org.mockito.junit.jupiter.MockitoExtension;
 
 import static com.watermelon.server.Constants.TEST_NOT_UID;
-import static com.watermelon.server.Constants.TEST_UID;
 import static org.junit.jupiter.api.Assertions.*;
 
 @ExtendWith(MockitoExtension.class)
