+++ src/test/java/com/watermelon/server/lottery/repository/LotteryApplierRepositoryTest.java
@@ -11,7 +11,7 @@ import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import java.util.List;
 import java.util.Optional;
 
-import static com.watermelon.server.Constants.TEST_UID;
+import static com.watermelon.server.event.lottery.auth.service.TestTokenVerifier.TEST_UID;
 import static org.assertj.core.api.Assertions.assertThat;
 
 @DataJpaTest
